services/kitchen/requirements.txt
requests
sqlalchemy
alembic
psycopg2-binary
google-genai
python-dotenv
